added h2 dependency, added test, created application.properties for tests purposes
