+++ backend/src/test/resources/application.properties
@@ -1,0 +1,11 @@
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=password
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.show-sql=true
+
+jwt.secret="tempToken"
+jwt.expiration.token.time = 1000
+jwt.expiration.refreshToken.time = 600
