+++ pom.xml
@@ -105,6 +105,11 @@
                 <version>5.4.5</version>
                 <scope>test</scope>
             </dependency>
+            <dependency>
+                <groupId>com.h2database</groupId>
+                <artifactId>h2</artifactId>
+                <version>1.4.200</version>
+            </dependency>
         </dependencies>
     </dependencyManagement>
